Created basic form layout of all required inputs
Still need to allow photos to take in file and convert it to url and save url. Still need to see if I can change manufacutred year to scroll through years. Still need to see if I can modify tags to take in multiple tags and also maybe even autofinish?